Moving tests to an existing file
